pdf_create_from_svg/requirements.txt
python-swiftclient==3.10.1
qrcode==6.1
cairosvg==2.4.2
Jinja2==2.11.2
lxml==4.5.2
